+++ KnowUrSystem/Model/ExpectModel.cs
@@ -7,9 +7,9 @@ namespace KnowUrSystem.Model
 {
     public class ExpectModel
     {
-        public double CumulativeProbability { get; set; }
+        public decimal CumulativeProbability { get; set; }
 
-        public double Expectancy { get; set; }
+        public decimal Expectancy { get; set; }
 
         public int Count { get; set; }
     }
